+++ src/views/Index.jsx
@@ -1,60 +1,20 @@
 import React from "react";
-// node.js library that concatenates classes (strings)
-import classnames from "classnames";
-// javascipt plugin for creating charts
-import Chart from "chart.js";
-// react plugin used to create charts
-import { Line, Bar } from "react-chartjs-2";
+
 // reactstrap components
 import {
   Button,
   Card,
   CardHeader,
   CardBody,
-  NavItem,
-  NavLink,
-  Nav,
-  Progress,
-  Table,
   Container,
   Row,
-  Col
+  Col,
+  CardTitle
 } from "reactstrap";
-
 // core components
-import {
-  chartOptions,
-  parseOptions,
-  chartExample1,
-  chartExample2
-} from "variables/charts.jsx";
-
 import Header from "components/Headers/Header.jsx";
 
-class Index extends React.Component {
-  state = {
-    activeNav: 1,
-    chartExample1Data: "data1"
-  };
-  toggleNavs = (e, index) => {
-    e.preventDefault();
-    this.setState({
-      activeNav: index,
-      chartExample1Data:
-        this.state.chartExample1Data === "data1" ? "data2" : "data1"
-    });
-    let wow = () => {
-      console.log(this.state);
-    };
-    wow.bind(this);
-    setTimeout(() => wow(), 1000);
-    // this.chartReference.update();
-  };
-  componentWillMount() {
-    if (window.Chart) {
-      parseOptions(Chart, chartOptions());
-    }
-  }
+class Tables extends React.Component {
   render() {
     return (
       <>
@@ -62,224 +22,174 @@ class Index extends React.Component {
         {/* Page content */}
         <Container className="mt--7 mb-4" fluid>
           <Row>
-            <Col className="mb-5 mb-xl-0" xl="8">
-              <Card className="bg-gradient-default shadow">
-                <CardHeader className="bg-transparent">
-                  <Row className="align-items-center">
-                    <div className="col">
-                      <h6 className="text-uppercase text-light ls-1 mb-1">
-                        Overview
-                      </h6>
-                      <h2 className="text-white mb-0">Flammables</h2>
+            <Col sm="12" md="3">
+              <Col sm="12">
+                <Card className="card-stats mb-4 mb-xl-0">
+                  <CardBody>
+                    <Row>
+                      <div className="col">
+                        <CardTitle
+                          tag="h5"
+                          className="text-uppercase text-muted mb-0"
+                        >
+                          Sensor Reading
+                          </CardTitle>
+                        <span className="h2 font-weight-bold mb-0">
+                          0
+                          </span>
+                      </div>
+                      <Col className="col-auto">
+                        <div className="icon icon-shape bg-success text-white rounded-circle shadow">
+                          <i className="fas fa-fire" />
+                        </div>
+                      </Col>
+                    </Row>
+                    <p className="mt-3 mb-0 text-muted text-sm">
+                      <span className="text-success mr-2">
+                        <i className="fa fa-arrow-up" /> 0%
+                        </span>{" "}
+                      <span className="text-nowrap">ppm</span>
+                    </p>
+                  </CardBody>
+                </Card>
+              </Col>
+
+              <Col sm="12">
+                <Card className="card-stats mb-4 mb-xl-0">
+                  <CardBody>
+                    <Row>
+                      <div className="col">
+                        <CardTitle
+                          tag="h5"
+                          className="text-uppercase text-muted mb-0"
+                        >
+                          Sensor Reading
+                          </CardTitle>
+                        <span className="h2 font-weight-bold mb-0">
+                          0
+                          </span>
+                      </div>
+                      <Col className="col-auto">
+                        <div className="icon icon-shape bg-success text-white rounded-circle shadow">
+                          <i className="fas fa-fire" />
+                        </div>
+                      </Col>
+                    </Row>
+                    <p className="mt-3 mb-0 text-muted text-sm">
+                      <span className="text-success mr-2">
+                        <i className="fa fa-arrow-up" /> 0%
+                        </span>{" "}
+                      <span className="text-nowrap">ppm</span>
+                    </p>
+                  </CardBody>
+                </Card>
+              </Col>
+            </Col>
+
+
+            <Col sm="12" lg="6">
+              <Card className="card-profile shadow">
+                <Row className="justify-content-center">
+                  <Col className="order-lg-2" lg="3">
+                    <div className="card-profile-image">
+
                     </div>
+                  </Col>
+                </Row>
+                <CardHeader className="text-center border-0 pt-8 pt-md-4 pb-0 pb-md-4">
+                  <div className="d-flex justify-content-between">
+
+                  </div>
+                </CardHeader>
+                <CardBody className="pt-0 pt-md-4">
+                  <Row>
                     <div className="col">
-                      <Nav className="justify-content-end" pills>
-                        <NavItem>
-                          <NavLink
-                            className={classnames("py-2 px-3", {
-                              active: this.state.activeNav === 1
-                            })}
-                            href="#pablo"
-                            onClick={e => this.toggleNavs(e, 1)}
-                          >
-                            <span className="d-none d-md-block">Month</span>
-                            <span className="d-md-none">M</span>
-                          </NavLink>
-                        </NavItem>
-                        <NavItem>
-                          <NavLink
-                            className={classnames("py-2 px-3", {
-                              active: this.state.activeNav === 2
-                            })}
-                            data-toggle="tab"
-                            href="#pablo"
-                            onClick={e => this.toggleNavs(e, 2)}
-                          >
-                            <span className="d-none d-md-block">Week</span>
-                            <span className="d-md-none">W</span>
-                          </NavLink>
-                        </NavItem>
-                      </Nav>
+                      <div className="card-profile-stats d-flex justify-content-center mt-md-5">
+                        <h1 className="text-success">You're Safe!</h1>
+                      </div>
                     </div>
                   </Row>
-                </CardHeader>
-                <CardBody>
-                  {/* Chart */}
-                  <div className="chart">
-                    <Line
-                      data={chartExample1[this.state.chartExample1Data]}
-                      options={chartExample1.options}
-                      getDatasetAtEvent={e => console.log(e)}
-                    />
+                  <div className="text-center">
+                    <h3>
+                      Ignio at a glance
+                    </h3>
+                    <div className="h5 font-weight-300">
+                      Your devices are working fine.
+                    </div>
+
+                    <hr className="my-4" />
+                    <p>
+                      Ignio is taking care of you.
+                    </p>
+                    <a href="#pablo" onClick={e => e.preventDefault()}>
+                      Show more
+                    </a>
                   </div>
                 </CardBody>
               </Card>
             </Col>
-            <Col xl="4">
-              <Card className="shadow">
-                <CardHeader className="bg-transparent">
-                  <Row className="align-items-center">
-                    <div className="col">
-                      <h6 className="text-uppercase text-muted ls-1 mb-1">
-                        Performance
-                      </h6>
-                      <h2 className="mb-0">Device Activity</h2>
-                    </div>
-                  </Row>
-                </CardHeader>
-                <CardBody>
-                  {/* Chart */}
-                  <div className="chart">
-                    <Bar
-                      data={chartExample2.data}
-                      options={chartExample2.options}
-                    />
-                  </div>
-                </CardBody>
-              </Card>
-            </Col>
-          </Row>
-          <Row className="mt-5">
-            <Col className="mb-5 mb-xl-0" xl="8">
-              <Card className="shadow">
-                <CardHeader className="border-0">
-                  <Row className="align-items-center">
-                    <div className="col">
-                      <h3 className="mb-0">Sensor Readings</h3>
-                    </div>
-                    <div className="col text-right">
-                      <Button
-                        color="primary"
-                        href="#pablo"
-                        onClick={e => e.preventDefault()}
-                        size="sm"
-                      >
-                        See all
-                      </Button>
-                    </div>
-                  </Row>
-                </CardHeader>
-                <Table className="align-items-center table-flush" responsive>
-                  <thead className="thead-light">
-                    <tr>
-                      <th scope="col">parameter</th>
-                      <th scope="col">Last Synchronization</th>
-                      <th scope="col">Percentage</th>
-                      <th scope="col">variation rate</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr>
-                      <th scope="row">Carbon Monoxide Gas</th>
-                      <td>31 Jul 2019 12:40 PM</td>
-                      <td>3.45%</td>
-                      <td>
-                        <i className="fas fa-arrow-up text-danger mr-3" />{" "}
-                        0.17%
-                      </td>
-                    </tr>
-                    <tr>
-                      <th scope="row">Temperature</th>
-                      <td>31 Jul 2019 12:40 PM</td>
-                      <td>3.45%</td>
-                      <td>
-                        <i className="fas fa-arrow-up text-warning mr-3" />{" "}
-                        1.23%
-                      </td>
-                    </tr>
-                    <tr>
-                      <th scope="row">LP Gas</th>
-                      <td>31 Jul 2019 12:40 PM</td>
-                      <td>6.85%</td>
-                      <td>
-                        <i className="fas fa-arrow-down text-success mr-3" />{" "}
-                        5.12%
-                      </td>
-                    </tr>
-                    <tr>
-                      <th scope="row">Other</th>
-                      <td>31 Jul 2019 12:40 PM</td>
-                      <td>0.91%</td>
-                      <td>
-                        <i className="fas fa-arrow-up text-primary mr-3" />{" "}
-                        3.18%
-                      </td>
-                    </tr>
-                  </tbody>
-                </Table>
-              </Card>
-            </Col>
-            <Col xl="4">
-              <Card className="shadow">
-                <CardHeader className="border-0">
-                  <Row className="align-items-center">
-                    <div className="col">
-                      <h3 className="mb-0">Alerts</h3>
-                    </div>
-                    <div className="col text-right">
-                      <Button
-                        color="primary"
-                        href="#pablo"
-                        onClick={e => e.preventDefault()}
-                        size="sm"
-                      >
-                        See all
-                      </Button>
-                    </div>
-                  </Row>
-                </CardHeader>
-                <Table className="align-items-center table-flush" responsive>
-                  <thead className="thead-light">
-                    <tr>
-                      <th scope="col">Incident</th>
-                      <th scope="col">Danger Level</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr>
-                      <th scope="row">Node test</th>
-                      <td>
-                        <div className="d-flex align-items-center">
-                          <span className="mr-2">0%</span>
-                          <div>
-                            <Progress
-                              max="100"
-                              value="0"
-                              barClassName="bg-gradient-danger"
-                            />
-                          </div>
+            <Col sm="12" md="3">
+              <Col sm="12">
+                <Card className="card-stats mb-4 mb-xl-0">
+                  <CardBody>
+                    <Row>
+                      <div className="col">
+                        <CardTitle
+                          tag="h5"
+                          className="text-uppercase text-muted mb-0"
+                        >
+                          Sensor Reading
+                          </CardTitle>
+                        <span className="h2 font-weight-bold mb-0">
+                          0
+                          </span>
+                      </div>
+                      <Col className="col-auto">
+                        <div className="icon icon-shape bg-success text-white rounded-circle shadow">
+                          <i className="fas fa-fire" />
                         </div>
-                      </td>
-                    </tr>
-                    <tr>
-                      <th scope="row">Minor smoke</th>
-                      <td>
-                        <div className="d-flex align-items-center">
-                          <span className="mr-2">20%</span>
-                          <div>
-                            <Progress
-                              max="100"
-                              value="20"
-                              barClassName="bg-gradient-success"
-                            />
-                          </div>
+                      </Col>
+                    </Row>
+                    <p className="mt-3 mb-0 text-muted text-sm">
+                      <span className="text-success mr-2">
+                        <i className="fa fa-arrow-up" /> 0%
+                        </span>{" "}
+                      <span className="text-nowrap">ppm</span>
+                    </p>
+                  </CardBody>
+                </Card>
+              </Col>
+
+              <Col sm="12">
+                <Card className="card-stats mb-4 mb-xl-0">
+                  <CardBody>
+                    <Row>
+                      <div className="col">
+                        <CardTitle
+                          tag="h5"
+                          className="text-uppercase text-muted mb-0"
+                        >
+                          Sensor Reading
+                          </CardTitle>
+                        <span className="h2 font-weight-bold mb-0">
+                          0
+                          </span>
+                      </div>
+                      <Col className="col-auto">
+                        <div className="icon icon-shape bg-success text-white rounded-circle shadow">
+                          <i className="fas fa-fire" />
                         </div>
-                      </td>
-                    </tr>
-                    <tr>
-                      <th scope="row">Major smoke</th>
-                      <td>
-                        <div className="d-flex align-items-center">
-                          <span className="mr-2">50%</span>
-                          <div>
-                            <Progress max="100" value="50" />
-                          </div>
-                        </div>
-                      </td>
-                    </tr>
-                  </tbody>
-                </Table>
-              </Card>
+                      </Col>
+                    </Row>
+                    <p className="mt-3 mb-0 text-muted text-sm">
+                      <span className="text-success mr-2">
+                        <i className="fa fa-arrow-up" /> 0%
+                        </span>{" "}
+                      <span className="text-nowrap">ppm</span>
+                    </p>
+                  </CardBody>
+                </Card>
+              </Col>
             </Col>
           </Row>
         </Container>
@@ -288,4 +198,4 @@ class Index extends React.Component {
   }
 }
 
-export default Index;
+export default Tables;
